use int32_t, size_t more

+++ sperl_sub.h
@@ -12,11 +12,11 @@ struct SPerl_sub {
   SPerl_ARRAY* op_my_vars;
   void (*native_address)(SPerl* sperl);
   const char* abs_name;
-  int64_t bytecode_base;
-  int64_t bytecode_length;
-  int64_t operand_stack_max;
-  int64_t constant_pool_address;
-  int64_t abs_name_constant_pool_address;
+  size_t bytecode_base;
+  size_t bytecode_length;
+  size_t operand_stack_max;
+  size_t constant_pool_address;
+  size_t abs_name_constant_pool_address;
   _Bool is_native;
   _Bool is_constant;
 };
